ParanoidOTA: GApps download API
Some time ago the Goo.im server was down, since that time GApps were not
uploaded to Goo.im any more and the update API didn't works.
Because of this fact I have searched for another API, but i haven't
found any well-functioning API. So I writed my own API for
BasketBuild.com.
I will take care of the PHP-API on the server.
I hope the patch will be in the next release ;-).

Signed-off-by: Matthias Freund <matthias88freund@gmail.com>
Signed-off-by: Ayysir <ayysir@aospal.com>

Change-Id: I3aa9f7cdebf8bc8b67f30fd36430978965cf0ee0